massto/rubygems.org

+++ app/jobs/delete_user.rb
@@ -6,7 +6,10 @@ class DeleteUser
   end
 
   def perform
-    profile = User.find(user["id"])
+    profile = User.find_by(id: user["id"])
+
+    return unless profile
+
     email = profile.email
     if profile.destroy
       Mailer.deletion_complete(email)
